+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 - Deprecate the extension name with `Extension` suffix, e.g. `FootnoteExtension` -> `Footnote`.
   And the old names will be removed by `v1.0.0`.
+- Store extra info after the language text in fenced code:
+  ~~~
+  ```python myscript.py    <-- myscript.py is stored in element
+  print('hello world')
+  ```
+  ~~~
+- Built-in code highlight extension using pygments.
 
 ## v0.6.0(2019-7-26)